chore(vrf): Add types/schema
